pyproject.toml
[project]
name = "neurality-health-voice-agent"
version = "1.0.0"
description = "Real-time voice agent using Twilio + LiveKit Agents + MCP tools"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=0.8.0",
    "livekit-plugins-deepgram>=0.6.0",
    "livekit-plugins-openai>=0.8.0",
    "livekit-plugins-silero>=0.6.0",
    "livekit-api>=0.6.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "twilio>=9.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
